+++ src/nodes/NGSI/encode/encode.js
@@ -1,0 +1,57 @@
+/*
+   MIT License
+
+   Copyright 2023 Kazuhito Suda
+
+   This file is part of node-red-contrib-letsfiware-NGSI
+
+   https://github.com/lets-fiware/node-red-contrib-letsfiware-NGSI
+
+   Permission is hereby granted, free of charge, to any person obtaining a copy
+   of this software and associated documentation files (the "Software"), to deal
+   in the Software without restriction, including without limitation the rights
+   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+   copies of the Software, and to permit persons to whom the Software is
+   furnished to do so, subject to the following conditions:
+
+   The above copyright notice and this permission notice shall be included in all
+   copies or substantial portions of the Software.
+
+   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+   AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+   SOFTWARE.
+ */
+
+'use strict';
+
+const lib = require('../../../lib.js');
+
+module.exports = function (RED) {
+  function encodeNode(config) {
+    RED.nodes.createNode(this, config);
+    const node = this;
+
+    node.on('input', async function (msg) {
+      switch (typeof msg.payload) {
+        case 'string':
+          msg.payload = lib.encodeforbiddenChar(msg.payload);
+          msg.statusCode = 200;
+          break;
+        case 'object':
+          msg.payload = lib.encodeNGSI(msg.payload, true);
+          msg.statusCode = 200;
+          break;
+        default:
+          msg.payload = { error: 'payload not string, JSON Object or JSON Array' };
+          msg.statusCode = 500;
+          break;
+      }
+      node.send(msg);
+    });
+  }
+  RED.nodes.registerType('NGSI Encode', encodeNode);
+};