+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## node-red-contrib-letsfiware-NGSI v0.14.0-next
 
+-  ADD encode / decode node (#133)
+
 ## node-red-contrib-letsfiware-NGSI v0.14.0 - 02 May, 2023
 
 -  ADD tutorial (#129)